Update DefaultPublicationWorkflow.java
Publish did not copy Attachments to the new Document and give an Exeption on the first try. Second try let publish the Document but without Attachments. 

With this, it worked for me, Attachments were in the new Dokument after Publishing, no more exeptions. Its just copy/past from copyContentsToNewVersion, with minor changes. Im not a Coder at all, my first try in Java.